Fixed background location updates

+++ app/src/main/java/com/norasoderlund/ridetrackerapp/RecorderElapsedSecondsEvent.kt
@@ -1,3 +1,3 @@
 package com.norasoderlund.ridetrackerapp
 
-data class RecorderElapsedSecondsEvent(val elapsedSeconds: Int, val formattedElapsedSeconds: String)
+data class RecorderElapsedSecondsEvent(val elapsedSeconds: Long, val formattedElapsedSeconds: String)
